Move mesh-half space intersection implementaiton into .cc file
This simply adds a .cc file and moves the implementaiton details into
that file (with appropriate template pre-declarations for the exercised
scalar types).

(Incidentally cleaned up the dependency; :posed_half_space
instead of :mesh_intersection.)
