+++ geometry/proximity/BUILD.bazel
@@ -305,11 +305,12 @@ drake_cc_library(
 
 drake_cc_library(
     name = "mesh_half_space_intersection",
+    srcs = ["mesh_half_space_intersection.cc"],
     hdrs = [
         "mesh_half_space_intersection.h",
     ],
     deps = [
-        ":mesh_intersection",
+        ":posed_half_space",
         ":surface_mesh",
         "//common",
     ],
